backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# RAG
langchain>=0.2.0
langchain-openai>=0.1.0
openai>=1.30.0

# Embeddings
sentence-transformers>=3.0.0
torch>=2.2.0

# Vector store
chromadb>=0.5.0

# Voice
openai-whisper>=20231117
azure-cognitiveservices-speech>=1.38.0

# Leads
gspread>=6.1.0
google-auth>=2.29.0

# Utils
python-dotenv>=1.0.0
langdetect>=1.0.9
